format raycasting init functions
